+++ R/auk-sampling.r
@@ -90,6 +90,7 @@ auk_sampling <- function(file, sep = "\t") {
         state = character(),
         county = character(),
         bbox = numeric(),
+        year = integer(),
         date = character(),
         time = character(),
         last_edited = character(),
@@ -158,6 +159,16 @@ print.auk_sampling <- function(x, ...) {
   } else {
     cat(paste0("Lon ", e[1], " - ", e[3], "; "))
     cat(paste0("Lat ", e[2], " - ", e[4]))
+  }
+  cat("\n")
+  # year filter
+  cat("  Years: ")
+  if (length(x$filters$year) == 0) {
+    cat("all")
+  } else if (length(x$filters$year) <= 10) {
+    cat(paste(x$filters$year, collapse = ", "))
+  } else {
+    cat(paste0(length(x$filters$year), " years"))
   }
   cat("\n")
   # date filter